replaced use of robotide.run.process.Process with wx.Process. The former was sometimes hanging for us (maybe due to a blocking read()?). Also, the former combined stdout and stderr and it's important to keep them distinct.
The one downside to this change is that we must create the command as
a string rather than a list of arguments, which means we could run
into quoting problems with arguments that have spaces or quotes in
them. There is an attempt to do proper quoting for the common cases,
and I verified that the test runner can run a suite even if it has a
space in its name. Since most arguments are passed to pybot via an
argument file, this shouldn't pose much of a risk.

I chose to modify the plugin rather than to enhance RIDE's Process
class out of fear I might introduce a regression in the other plugins
that uses that class.