variablelist: Focus is now kept when moving vars
The synchronization of variables between the tree and the list view caused
the focus to transfer sometimes to the tree. If the move up/down is initiated
from list, we now grab the focus back always.

Additionally fixed a traceback in Linux that came when variables were moved
from the list view without their parent being open in the tree.