Show warning when wxPython < 2.8.12.1
Warning is shown always when RIDE is started. This ought to annoy users enough to make them upgrade or tell us that older versions really needs to be supported.

The committed version shows warning always to make it easier for others to verify how the dialog and its test. I've already checked that it looks OK on Linux (Mint) and Win7.

This fixes http://code.google.com/p/robotframework-ride/issues/detail?id=1141