Small fix for read-only installations
When RIDE is installed in read-only locations users are faced with this
the first time they run it:

 IOError: [Errno 13] Permission denied: '/home/bfo/.robotframework/ride/settings.cfg'

This happens because shutil.copy copies the read-only permission bits
from settings.cfg in the install path over to the users home directory.
By using shutil.copyfile instead, like this patch does, only the file
contents is copied and not the permission bits, so the file ends up
being writeable for the user.

This patch is specifically needed for the Nix package system where
applications are "locked down" in read-only paths.