Added "Run with Debug" hotkey  F9
I wanted to add a hotkey so I can run the currently selected test with
Log Level Debug
The simplest way was to copy the test runner plugin.
I added a bug  icon as start button from famfamfam as some the other
icon are used.
I'm not sure but the green color breaks the look and feel a little.