Update treenodehandlers.py to prevent freeze ups (#2205)
The _ActionHandler class is based off wx.Window which causes a native widget to be created. Avoid adding this native window to the tree to prevent Windows OS freezes when the tree is scrolled. Add it to the tree's parent instead.
Looks like the only wx.Window feature utilized is for popup menus when a tree item is right-clicked.
Attempts to fix #1843.